Arguments of optional closure types without `@escaping` are also warned
