+++ Sources/WeakSelfCheckCore/Extension/FunctionParameterSyntax+.swift
@@ -15,6 +15,11 @@ extension FunctionParameterSyntax {
         type.isFunctionType
     }
 
+    /// A boolean value that indicates whatever the parameter is a optional type.
+    var isOptionalType: Bool {
+        type.isOptionalType
+    }
+
     /// A boolean value that indicates whether the parameter has `@escaping` attribute.
     var isEscaping: Bool {
         guard let type = type.as(AttributedTypeSyntax.self) else {
@@ -38,41 +43,5 @@ extension AttributeSyntax {
             return false
         }
         return identifierType.name.tokenKind == .identifier("escaping")
-    }
-}
-
-extension TypeSyntaxProtocol {
-    /// A boolean value that indicates whatever the type is a closure type.
-    var isFunctionType: Bool {
-        if kind == .functionType {
-            return true
-        }
-
-        if let type = self.as(MemberTypeSyntax.self) {
-            if type.name.tokenKind == .identifier("Optional"),
-               let genericArgumentClause = type.genericArgumentClause,
-               let argument = genericArgumentClause.arguments.first {
-                return argument.argument.isFunctionType
-            }
-            return type.baseType.isFunctionType
-        }
-        if let type  = self.as(AttributedTypeSyntax.self) {
-            return type.baseType.isFunctionType
-        }
-        if let type = self.as(OptionalTypeSyntax.self) {
-            return type.wrappedType.isFunctionType
-        }
-        if let type = self.as(ImplicitlyUnwrappedOptionalTypeSyntax.self) {
-            return type.wrappedType.isFunctionType
-        }
-
-        if let type = self.as(IdentifierTypeSyntax.self),
-           type.name.tokenKind == .identifier("Optional"),
-           let genericArgumentClause = type.genericArgumentClause,
-           let argument = genericArgumentClause.arguments.first {
-            return argument.argument.isFunctionType
-        }
-
-        return false
     }
 }
